% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flipacoin.R
\name{Flipacoin}
\alias{Flipacoin}
\title{A Flipacoin Function}
\usage{
Flipacoin(nb = 1)
}
\description{
This function takes as an argument nb, the number of tosses, and give back x, 
a series of heads of tails.
}
\examples{
Flipacoin()
}

